+++ app/controllers/client_project_controller.rb
@@ -1,7 +1,7 @@
 class ClientProjectController < ApplicationController
   before_action :authenticate_client!
   before_action :set_project
-  before_action :invert,              only: [ :script, :timeline, :casting ]
+  before_action :invert,              only: [ :timeline, :casting, :category, :document ]
 
   def root
     render 'menu'
@@ -9,6 +9,36 @@ class ClientProjectController < ApplicationController
 
   def menu
 
+  end
+
+  def category
+    @category = @project.categories.find_by_param(params[:category_id])
+    if @category
+      @document = @category.pdf_documents.where(archived: false).first
+    else
+      redirect_to client_project_root_path(@project)
+    end
+  end
+
+  def document
+    @category = @project.categories.find_by_param(params[:category_id])
+    if @category
+      @document = @project.documents.find_by_id(params[:document_id])
+      render 'category'
+    else
+      redirect_to client_project_root_path(@project)
+    end
+  end
+
+  def download_document
+    @document = @project.documents.find_by_id(params[:document_id])
+    if @document and @document.document.size > 0
+      disposition = (params[:inline] == '1' ? 'inline' : 'attachment')
+      type = (@document.pdf? ? 'application/pdf' : 'application/octet-stream')
+      send_file File.join( CarrierWave::Uploader::Base.root, @document.document.url ), type: type, disposition: disposition
+    else
+      head :ok
+    end
   end
 
   private