+++ config/routes.rb
@@ -3,10 +3,16 @@ Rails.application.routes.draw do
   resources :projects do
     member do
       get :menu
+      get :logo
     end
 
     resources :categories
 
+    resources :documents do
+      member do
+        get :download
+      end
+    end
   end
 
   scope module: 'welcome' do
@@ -37,11 +43,14 @@ Rails.application.routes.draw do
   scope module: 'client_project' do
     get ":id", action: 'root', as: :client_project_root
     get ":id/menu", action: 'menu', as: :client_project_menu
-    get ":id/creative/script", action: 'script', as: :client_project_script
     get ":id/production/casting", action: 'casting', as: :client_project_casting
     get ":id/timeline", action: 'timeline', as: :client_project_timeline
 
-    get ":id/script", action: 'script'
+    get ":id/creative/:category_id", action: 'category', as: :client_project_category
+    get ":id/creative/:category_id/:document_id", action: 'document', as: :client_project_category_document
+    get ":id/documents/download/:document_id", action: 'download_document', as: :client_project_download_document
+    get ":id/:category_id", action: 'category'
+
     get ":id/casting", action: 'casting'
   end