+++ CHANGELOG.md
@@ -15,6 +15,8 @@
     - Client Username
     - Client Password
   - System admins can now create and edit project categories
+  - System admins can now upload documents to categories
+    - Documents can be archived
 - **General Changes**
   - Added a version page
   - Updated font location for better asset management using Rails